responsivenes

+++ lms/src/components/sideBar/SideBar.js
@@ -6,19 +6,41 @@ import { VscHome } from "react-icons/vsc";
 import { FiLogIn } from "react-icons/fi";
 import styles from "./SideBar.module.css"
 import Logo from "../images/LogoWhite.png"; 
-import { NavLink, Link, useNavigate } from "react-router-dom";
+import { NavLink, Link } from "react-router-dom";
 import { AiOutlinePieChart } from 'react-icons/ai';
+import axios from "axios";
 
 
 
 function SideBar(){
-    const navigate = useNavigate();
- 
-    const handleLogout = () => {
-        localStorage.clear();
 
-        navigate('/')
-    }
+ 
+    const handleLogout = async () => {
+        //   e.preventDefault();
+        let token = localStorage.getItem("token")
+        let email = localStorage.getItem("userEmail")
+          const data = {
+            email: email,
+          }
+          const url = "https://kidtots.onrender.com/student/logout"
+        
+          await axios.post(url, data, {
+            headers:{
+            "Authorization": "Bearer" + token,
+            "Content-Type": "application/json",
+            }
+           }).then((response) =>{
+         
+            if(response.status === 200){
+                console.log(response)
+                localStorage.clear();
+            // setEmail("");
+            }
+           }).then((error) =>{
+
+           })     
+        }   
+  
 
 return(
     <>
@@ -89,7 +111,7 @@ return(
                 <div className={styles.sideBarFooter}>
                     <div>
                         <Link to="/">
-                        <div>
+                        <div onClick={handleLogout}>
                         <i><FiLogIn /></i><p>Logout</p>
                         </div>
                         </Link>
